requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
frozenlist==1.8.0
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.1
numpy==2.1.2
pillow==11.0.0
propcache==0.4.1
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.2.1
requests==2.32.3
sympy==1.13.3
torch==2.1.2+cu121
torchvision==0.16.2+cu121
typing-inspection==0.4.2
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.22.0
